Fix PlatformChartRankComparisonLineChart
